Install with setuptools for dependency management
Setuptools lets you specify a `install_requires` block which specifies
dependencies. If you install with pip, or have setuptools installed
then we will install requests if needed. Otherwise issue a warning.
It looks like virtualenv comes with setuptools already, so this should
be a good change.

Also move the extractor from scripts to additional data (we still need
to decide if having it here is the right idea). Fixes #26